[INWA-506 - SplashScreen Compatibility]
-added landscape xml-files
-MainActivityViewModel now saves the current fragment
-added isInitliazied checks in CompanionMapFragment.kt
-added landscape compatibility for the CompanionMap (TODO: change xml file)

+++ app/src/main/java/de/infoware/followmesdkexample/companionmap/MapControlsFragment.kt
@@ -72,7 +72,7 @@ class MapControlsFragment : Fragment() {
         }
 
         // Observer for the crossingInfoReceived Callback: metersToCrossing
-        val metersToCrossingObserver = Observer<Double> { metersToCrossing ->
+        val metersToCrossingObserver = Observer<Int> { metersToCrossing ->
             tvMetersToCrossing.text = "$metersToCrossing m"
         }
